feat: refactored radarsvg in seperate component
